+++ packages/elements/src/tree/helpers/types.ts
@@ -7,10 +7,12 @@ export interface TreeDataItem extends DataItem {
    * Icon to show, when rendering the item.
    */
   icon?: string;
+
   /**
    * Label to show, when rendering the item.
    */
   label?: string;
+
   /**
    * Expanded state of child items.
    * If `true`, child items will be visible